backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

anthropic==0.34.2
supabase==2.7.4
httpx==0.27.2

python-pptx==1.0.2
PyMuPDF==1.24.10
mammoth==1.8.0
openpyxl==3.1.5

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

pytest==8.3.3
pytest-asyncio==0.24.0
